Vote calculation implemented.

+++ src/main/java/lt/vaidotas/food/business/voting/model/VoteResult.java
@@ -5,9 +5,9 @@ import lt.vaidotas.food.business.restaurant.model.Restaurant;
 public class VoteResult {
 
     private final Restaurant restaurant;
-    private final Integer count;
+    private final Long count;
 
-    public VoteResult(final Restaurant restaurant, final Integer count) {
+    public VoteResult(final Restaurant restaurant, final Long count) {
         this.restaurant = restaurant;
         this.count = count;
     }
@@ -16,7 +16,7 @@ public class VoteResult {
         return restaurant;
     }
 
-    public Integer getCount() {
+    public Long getCount() {
         return count;
     }
 }
